1解决valueText值转出后是以科学计数法格式输出问题（1e-7） 2优化internalTxn数据处理，新增字段init、address、code、type 3获取账户余额存储到风险名单的balance字段中 4解决添加风险名单来源信息bug 5优化erc20Txn,logs中地址显示格式

+++ services/bitcoin/bitcoin.go
@@ -130,7 +130,7 @@ func GetAddressInfo(bitClient *BitClient, addr string) (float64, error) {
 	if err != nil {
 		return 0, fmt.Errorf("get balance Error:%v", err)
 	}
-	balanceFloat, _ := ConvertSatoshiToBTC(big.NewInt(int64(balance)))
+	balanceFloat, _ := ConvertSatoshiToBTC(big.NewInt(balance))
 	return balanceFloat, nil
 
 }
